conditionally use afterAll

+++ libs/res-to-ts/src/convert.test.ts
@@ -26,11 +26,13 @@ test('creating a resource for a directory', async () => {
 
     const copiedDirectories: string[] = [];
 
-    afterAll(() => {
-      for (const copiedDirectory of copiedDirectories) {
-        fs.rmSync(copiedDirectory, { recursive: true, force: true });
-      }
-    });
+    if (typeof jest !== 'undefined') {
+      afterAll(() => {
+        for (const copiedDirectory of copiedDirectories) {
+          fs.rmSync(copiedDirectory, { recursive: true, force: true });
+        }
+      });
+    }
 
     export function asDirectoryPath(): string {
       const tmpDir = fs.mkdtempSync(tmpdir() + sep);
